Change: Delete result button working but app unstable

+++ src/se/team05/content/Result.java
@@ -32,7 +32,7 @@ public class Result {
 	
 	private int _id;
 	private long routId;
-	private int timestamp;
+	private long timestamp;
 	private int time;
 	private int calories;
 	private int distance;
@@ -67,7 +67,7 @@ public class Result {
 	 * @param calories
 	 * 		calories (kcal) used during route
 	 */
-	public Result(int _id, long routId, int timestamp, int time, int distance, int calories)
+	public Result(int _id, long routId, long timestamp, int time, int distance, int calories)
 	{
 		this._id = _id;
 		this.routId = routId;
@@ -92,7 +92,7 @@ public class Result {
 	 * @param calories
 	 * 		calories (kcal) used during route
 	 */
-	public Result(long routId, int timestamp, int time, int distance, int calories)
+	public Result(long routId, long timestamp, int time, int distance, int calories)
 	{
 		this.routId = routId;
 		this.timestamp = timestamp;
@@ -123,7 +123,7 @@ public class Result {
 	 * @return time and date when the result was created 
 	 * 			in the format of nanoseconds passed since year 1970
 	 */
-	public int getTimestamp() {
+	public long getTimestamp() {
 		return timestamp;
 	}
 	/**
@@ -133,7 +133,7 @@ public class Result {
 	 * 		time and date when the result is created
 	 * 		in the format of nanoseconds passed since year 1970
 	 */
-	public void setTimestamp(int timestamp) {
+	public void setTimestamp(long timestamp) {
 		this.timestamp = timestamp;
 	}
